Add taproot key path test vector
